Functions

Function Reference

Function Reference

Common

Block Matching Feature Tracking

KLT Feature Tracking

Camera Models and Geometry

Inertial Sensor Aided Feature Tracking

Index

Common

SensorFeatureTracking.drawfeatureLine! — Function.

drawfeatureLine!(image,feature1, feature2)

Draw a line between 2 features.

source

SensorFeatureTracking.drawfeatureX! — Function.

drawfeatureX!(image,feature [,crosslength = 2])

Draw a + on a feature.

Examples

julia> map(ft->drawfeatureX!(image, ft, length=5),features)

source

SensorFeatureTracking.getApproxBestHarrisInWindow — Function.

getApproxBestHarrisInWindow(image,[nfeatures=100, window = 9, k=0.04, stepguess=0.9, threshold = 1e-4])

Return the n best Harris features in a window

source

SensorFeatureTracking.getApproxBestShiTomasi — Function.

getApproxBestShiTomasi(image,[n=100, windowSize = 9, k=0.04, stepguess=0.4, threshold = 1e-4])

Return the n aproxamate best Shi Tomasi features in a window

source

Block Matching Feature Tracking

SensorFeatureTracking.compute_sad — Function.

compute_sad(image1, image2, off1X, off1Y, off2X, off2Y, REGION_SIZE)

Compute Sum of Absolute Differences of two regions with size REGION_SIZE.

source

SensorFeatureTracking.compute_ssd — Function.

compute_ssd(image1, image2, off1X, off1Y, off2X, off2Y, REGION_SIZE)

Compute Sum of Squared Differences of two regions with REGION_SIZE.

source

SensorFeatureTracking.compute_ncc — Function.

compute_ncc(image1, image2, off1X, off1Y, off2X, off2Y, REGION_SIZE)

Compute the normalized cross-correlation of two regions with size REGION_SIZE.

source

SensorFeatureTracking.block_tracker! — Function.

block_tracker(tracker::BlockTracker, image1, image2)

Track features between two images.

source

KLT Feature Tracking

Forward KLT tracking with pyramids

SensorFeatureTracking.ImageTrackerSetup — Function.

ImageTrackerSetup(orgI_setup, corners,[windowSize = 20, downsampleFactor_setup = 2])

orgI_setup: Image to do trackingon corners: List of corners, preferably a returned list from getApproxBestHarrisInWindow() windowSize Half the size of the window to look for a matching feature. Actual window size is windowSize * 2 + 1, windowSize = 20 will result in a 41x41 window downsampleFactor_setup downsampleFactor_setup = 2: use pyramid images (first track on a downsampled image, then on the regular image with a smaller windowsize), increases speed. Set downsampleFactor_setup = 1 to use only regular image

Returns two structures, one of type ImageTrackerVariables and one of type ImageTrackerConstants

source

SensorFeatureTracking.trackOneFeature — Function.

trackOneFeature(ITVar, ITConst, corner_i)

ITVar: construct of variables of type ImageTrackerVariables ITConst: construct of constants of type ImageTrackerConstants corner_i:index of selected corner in ITVar.p_reference to do tracking on

Returns nothing, updates p_reference[2,corner_i] with the new coordinates of the tracked feature Does not use orgI_downsample or I_nextFrame_downsample at all, downsampleFactor must be 1

source

SensorFeatureTracking.KTL_Tracker! — Function.

KTL_Tracker!(ITVar, ITConst)

ITVar: construct of variables of type ImageTrackerVariables ITConst: construct of constants of type ImageTrackerConstants

Returns nothing, updates p_reference[2,:] with the new coordinates of the tracked features

source

SensorFeatureTracking.trackOneFeaturePyramid — Function.

trackOneFeature(ITVar, ITConst, corner_i)

ITVar: construct of variables of type ImageTrackerVariables ITConst: construct of constants of type ImageTrackerConstants corner_i:index of selected corner in ITVar.p_reference to do tracking on

Returns nothing, updates p_reference[2,corner_i] with the new coordinates of the tracked feature Use orgI_downsample and I_nextFrame_downsample to track features, downsampleFactor must be 2 for now, can be increased in future work

source

SensorFeatureTracking.warpping! — Function.

warpping!(Iout,Iin,x,y,M)

Iout: returned warped image Iin: image to be warped x: derivatives kernel on x axis y: derivatives kernel on y axis M: affine matrix for template rotation and translation

Return a warped image

source

Camera Models and Geometry

SensorFeatureTracking.CameraModelandParameters — Type.

Data structure for a Camera model with parameters. Use CameraModel(width,height,fc,cc,skew,kc) for easy construction.

source

SensorFeatureTracking.CameraModel — Function.

CameraModel(width,height,fc,cc,skew,kc)

Constructor helper for creating a camera model.

source

Inertial Sensor Aided Feature Tracking

SensorFeatureTracking.IMU_DATA — Type.

Data structure for holding time, gyroscope, and accelerometer data.

Fields

utime – time [μs]

acc – accelerometer data

gyro – gyroscope data

source

SensorFeatureTracking.estimateRotationFromKeypoints — Function.

estimateRotationFromKeypoints(points_a, points_b)

Estimate the rotation between 2 sets of Keypoints a and be using HornAbsoluteOrientation

source

SensorFeatureTracking.predictHomographyIMU! — Function.

predictHomographyIMU!(index, current_time, vector_data, CameraK, CameraK_inverse)

Estimate camera rotation from IMU data and compute predicted homography

source

SensorFeatureTracking.integrateGyroBetweenFrames! — Function.

integrateGyroBetweenFrames!(index, current_time, vector_data)

Estimate rotations from IMU data between time stamps.

source

SensorFeatureTracking.HornAbsoluteOrientation — Function.

HornAbsoluteOrientation(⊽a, ⊽b)

Compute the rotation between rows of (a and b)::Array{Float64,2}. Rotate b into the frame of a Returns a quaternion, aQb

source

Index

SensorFeatureTracking.CameraModelandParameters

SensorFeatureTracking.IMU_DATA

SensorFeatureTracking.CameraModel

SensorFeatureTracking.HornAbsoluteOrientation

SensorFeatureTracking.ImageTrackerSetup

SensorFeatureTracking.KTL_Tracker!

SensorFeatureTracking.block_tracker!

SensorFeatureTracking.compute_ncc

SensorFeatureTracking.compute_sad

SensorFeatureTracking.compute_ssd

SensorFeatureTracking.drawfeatureLine!

SensorFeatureTracking.drawfeatureX!

SensorFeatureTracking.estimateRotationFromKeypoints

SensorFeatureTracking.getApproxBestHarrisInWindow

SensorFeatureTracking.getApproxBestShiTomasi

SensorFeatureTracking.integrateGyroBetweenFrames!

SensorFeatureTracking.predictHomographyIMU!

SensorFeatureTracking.trackOneFeature

SensorFeatureTracking.trackOneFeaturePyramid

SensorFeatureTracking.warpping!